// src/app/components/merchandising/Costing/model/costing-accessories.model.ts
export class CostAccessories {
    id:number;
    itemId: number;
    itemName: string;
    costMasterId: number;
    usedPlace: string;
    refCode: string;
    stCode: string;
    description: string;
    nominationStatus: string;
    nominationName: string;
    supplierId: number;
    supplierName: string;
    consumption:number;
    wastagePercentage: number;
    coloumColor: string;
    totalConsumption: number;
    rate:number;
    finCostPc: number;
    value: number;
    categoryName: string;
    uomId: number;
    uomName: string;
    mkuId: number;
    mkuName: string;
    marketRelation: number;
    countryId: number;
    countryName: string;
    categoryId: number;
    isFinCostPc: number;
    costCategoryGroup: string;
    updateFrom: string;
    accessoriesComment: string;
    isChange: string;
    costingStatus: string;
    texId: number;
    itemColor: string;
    modelCode: string;
    dsmCode: string;
    costItemSize: string;

    //new 2
    //pdmCode : string;
}